Create riggtravis.md
I have created a fork of Aseprite so I can do a few bug fixes in the fedora environment. I already messed up by not paying attention to the rules of how to pull requests are to be done, so I'm not sure if I'll be able to do the branching that was required. I might be able to fenagle it, but I just really want to help out.